feat: check middleware files
