+++ src/middleware/index.ts
@@ -1,3 +1,4 @@
-export * from './database';
-export * from './fieldsValidator';
-export * from './logger';
+export * from './utils';
+export * from './validators/database';
+export * from './validators/fields';
+export * from './validators';
