+++ src/middleware/utils/index.ts
@@ -1,10 +1,1 @@
-export const getFirstTruthyValue = (
-  ...values: Array<string | undefined | null>
-): string => {
-  for (const value of values) {
-    if (value) {
-      return String(value);
-    }
-  }
-  return '';
-};
+export * from './logger';
