+++ src/middleware/validators/fields/fieldsOfBody.ts
@@ -1,18 +1,6 @@
 import type { NextFunction, Request, Response } from 'express';
 
-import { badRequest } from '../common';
-
-const validatesRequiredParam = (
-  field: string,
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Response | undefined => {
-  if (!field || !req?.params?.[field]) {
-    return badRequest(`check '${field}' field.`, res);
-  }
-  next();
-};
+import { badRequest } from '../../../common';
 
 const validatesRequiredFieldOfBody = (
   field: string,
@@ -25,50 +13,6 @@ const validatesRequiredFieldOfBody = (
   }
   next();
 };
-
-const validatesTypeOfFieldNumber = (
-  field: string,
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Response | undefined => {
-  if (isNaN(Number(req?.params?.[field]))) {
-    return badRequest(`field '${field}' only accepts numbers.`, res);
-  }
-  next();
-};
-
-export const validatesHotelCodeParam = (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Response | undefined => validatesRequiredParam('hotelCode', req, res, next);
-
-export const validatesRoomCodeParam = (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Response | undefined => validatesRequiredParam('roomCode', req, res, next);
-
-export const validatesRateCodeParam = (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Response | undefined => validatesRequiredParam('rateCode', req, res, next);
-
-export const validatesInventoryIdParam = (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Response | undefined =>
-  validatesRequiredParam('inventoryId', req, res, next);
-
-export const validatesInventoryIdType = (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Response | undefined =>
-  validatesTypeOfFieldNumber('inventoryId', req, res, next);
 
 export const validatesNameFieldOfBody = (
   req: Request,
