+++ src/middleware/validators/fields/index.ts
@@ -1,0 +1,3 @@
+export * from './fieldsOfBody';
+export * from './fieldsOfParams';
+export * from './typeOfParams';
